fix stack initialization issues
I was not initializing the stack appropriately, since the entrypoint is
not a `naked` function, which sets up it's own stack which is not what
we want.

This caused the `a0` and `a1` to be both zero, while we expect them to
be valid values, the latter of which contains the pointer to the device
tree.